chore: config: kickstart configfile processing
move configfile processing in its own file,
to be expanded later.

Internal-only change.
